requirements.txt
pyspark
requests
python-dotenv==0.21.0
pymongo
dbt-core==1.8.6
dbt-clickhouse==1.8.3
great_expectations==0.18.17
apache-airflow-client==2.9.0
apache-airflow-providers-apache-spark==4.8.1